+++ packages/eyes-webdriverio-4/test/it/spec-driver.spec.js
@@ -183,8 +183,8 @@ describe('spec driver', async () => {
   describe('native app (@mobile @native)', async () => {
     before(async () => {
       ;[browser, destroyBrowser] = await spec.build({
-        app: 'http://saucelabs.com/example_files/ContactManager.apk',
-        device: 'Android Emulator',
+        app: 'https://applitools.jfrog.io/artifactory/Examples/android/1.3/app-debug.apk',
+        device: 'Pixel 3a XL',
         orientation: 'landscape',
       })
     })
@@ -202,11 +202,14 @@ describe('spec driver', async () => {
     it('getDriverInfo()', async () => {
       await getDriverInfo({
         expected: {
-          deviceName: 'Android Emulator',
+          deviceName: 'Google Pixel 3a XL GoogleAPI Emulator',
           isMobile: true,
           isNative: true,
           platformName: 'Android',
-          platformVersion: '6.0',
+          platformVersion: '10.0',
+          pixelRatio: 2.5,
+          statusBarHeight: 60,
+          navigationBarHeight: 1080,
         },
       })
     })