+++ packages/screenshoter/src/take-viewport-screenshot.js
@@ -1,7 +1,7 @@
 const utils = require('@applitools/utils')
 const makeTakeScreenshot = require('./take-screenshot')
 
-async function takeViewportScreenshot({logger, context, region, wait, stabilization, debug = {}}) {
+async function takeViewportScreenshot({context, region, withStatusBar, wait, stabilization, debug = {}, logger}) {
   logger.verbose('Taking image of...')
 
   const driver = context.driver
@@ -9,12 +9,12 @@ async function takeViewportScreenshot({logger, context, region, wait, stabilizat
 
   await utils.general.sleep(wait)
 
-  const image = await takeScreenshot()
+  const image = await takeScreenshot({withStatusBar})
 
   if (region) {
     const cropRegion = await driver.getRegionInViewport(context, region)
     if (utils.geometry.isEmpty(cropRegion)) throw new Error('Screenshot region is out of viewport')
-    await image.crop(cropRegion)
+    image.crop(cropRegion)
     await image.debug({path: debug.path, suffix: 'region'})
     return {image, region: cropRegion}
   } else {