+++ packages/screenshoter/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## Unreleased
 
+- add `withStatusBar` capability property to take app and full app screenshots with status bar
+- fix ios web screenshots
 
 ## 3.2.1 - 2021/8/8