+++ packages/types/CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## Unreleased
 
-- add support for `throwErr` argument to `close` and `closeAllEyes`
-- return array of tests results from `close` and `abort`
+- add optional property `statusBarHeight` to the `DriverInfo` type
 
 ## 1.0.7 - 2021/8/13